requirements.txt
requests>=2.31,<3.0
pydantic>=2.5,<3.0
rich>=13.7,<14.0
python-dotenv>=1.0,<2.0
